(refactor) Put order_book_message and order_book_tracker_entry into different files
