+++ hummingbot/market/bittrex/bittrex_order_book.pyx
@@ -1,19 +1,19 @@
 #!/usr/bin/env python
-from aiokafka import ConsumerRecord
-import bz2
+
 import logging
-from sqlalchemy.engine import RowProxy
 from typing import (
-    Optional,
+    Any,
     Dict,
-    List, Any)
-import ujson
+    List,
+    Optional,
+)
 
 from hummingbot.logger import HummingbotLogger
-from hummingbot.core.event.events import TradeType
+from hummingbot.market.bittrex.bittrex_order_book_message import BittrexOrderBookMessage
 from hummingbot.core.data_type.order_book cimport OrderBook
 from hummingbot.core.data_type.order_book_message import (
-    OrderBookMessage, OrderBookMessageType, BittrexOrderBookMessage
+    OrderBookMessage,
+    OrderBookMessageType,
 )
 
 _btob_logger = None
