+++ hummingbot/market/bittrex/bittrex_order_book_tracker_entry.py
@@ -1,0 +1,23 @@
+from hummingbot.core.data_type.order_book import OrderBook
+from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
+from hummingbot.market.bittrex.bittrex_active_order_tracker import BittrexActiveOrderTracker
+
+
+class BittrexOrderBookTrackerEntry(OrderBookTrackerEntry):
+    def __init__(self,
+                 trading_pair: str,
+                 timestamp: float,
+                 order_book: OrderBook,
+                 active_order_tracker: BittrexActiveOrderTracker):
+        self._active_order_tracker = active_order_tracker
+        super(BittrexOrderBookTrackerEntry, self).__init__(trading_pair, timestamp, order_book)
+
+    def __repr__(self) -> str:
+        return (
+            f"BittrexOrderBookTrackerEntry(trading_pair='{self._trading_pair}', timestamp='{self._timestamp}', "
+            f"order_book='{self._order_book}')"
+        )
+
+    @property
+    def active_order_tracker(self) -> BittrexActiveOrderTracker:
+        return self._active_order_tracker
