+++ hummingbot/market/ddex/ddex_api_order_book_data_source.py
@@ -21,13 +21,13 @@ from hummingbot.core.utils import async_ttl_cache
 from hummingbot.core.utils.async_utils import safe_gather
 from hummingbot.market.ddex.ddex_active_order_tracker import DDEXActiveOrderTracker
 from hummingbot.market.ddex.ddex_order_book import DDEXOrderBook
+from hummingbot.market.ddex.ddex_order_book_message import DDEXOrderBookMessage
 from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_tracker_entry import (
     DDEXOrderBookTrackerEntry,
     OrderBookTrackerEntry
 )
-from hummingbot.core.data_type.order_book_message import DDEXOrderBookMessage
 
 TRADING_PAIR_FILTER = re.compile(r"(TUSD|WETH|DAI|SAI)$")
