+++ hummingbot/market/ddex/ddex_order_book_message.py
@@ -1,0 +1,71 @@
+#!/usr/bin/env python
+
+from typing import (
+    Dict,
+    List,
+    Optional,
+)
+
+from hummingbot.core.data_type.order_book_row import OrderBookRow
+from hummingbot.core.data_type.order_book_message import (
+    OrderBookMessage,
+    OrderBookMessageType,
+)
+
+
+class DDEXOrderBookMessage(OrderBookMessage):
+    def __new__(
+        cls,
+        message_type: OrderBookMessageType,
+        content: Dict[str, any],
+        timestamp: Optional[float] = None,
+        *args,
+        **kwargs,
+    ):
+        if timestamp is None:
+            if message_type is OrderBookMessageType.SNAPSHOT:
+                raise ValueError("timestamp must not be None when initializing snapshot messages.")
+            timestamp = content["time"] * 1e-3
+        return super(DDEXOrderBookMessage, cls).__new__(
+            cls, message_type, content, timestamp=timestamp, *args, **kwargs
+        )
+
+    @property
+    def update_id(self) -> int:
+        return int(self.timestamp * 1e3)
+
+    @property
+    def trade_id(self) -> int:
+        return int(self.timestamp * 1e3)
+
+    @property
+    def trading_pair(self) -> str:
+        return self.content["marketId"]
+
+    @property
+    def asks(self) -> List[OrderBookRow]:
+        raise NotImplementedError("DDEX order book messages have different semantics.")
+
+    @property
+    def bids(self) -> List[OrderBookRow]:
+        raise NotImplementedError("DDEX order book messages have different semantics.")
+
+    @property
+    def has_update_id(self) -> bool:
+        return True
+
+    @property
+    def has_trade_id(self) -> bool:
+        return True
+
+    def __eq__(self, other) -> bool:
+        return self.type == other.type and self.timestamp == other.timestamp
+
+    def __lt__(self, other) -> bool:
+        if self.timestamp != other.timestamp:
+            return self.timestamp < other.timestamp
+        else:
+            """
+            If timestamp is the same, the ordering is snapshot < diff < trade
+            """
+            return self.type.value < other.type.value
