+++ hummingbot/market/dolomite/dolomite_order_book.pxd
@@ -2,4 +2,3 @@ from hummingbot.core.data_type.order_book cimport OrderBook
 
 cdef class DolomiteOrderBook(OrderBook):
     pass
-
