+++ hummingbot/market/dolomite/dolomite_order_book_tracker.py
@@ -6,12 +6,12 @@ from hummingbot.market.dolomite.dolomite_active_order_tracker import DolomiteAct
 from hummingbot.logger import HummingbotLogger
 from hummingbot.core.data_type.order_book_tracker import OrderBookTracker, OrderBookTrackerDataSourceType
 from hummingbot.market.dolomite.dolomite_order_book import DolomiteOrderBook
+from hummingbot.market.dolomite.dolomite_order_book_message import DolomiteOrderBookMessage
 from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
 from hummingbot.core.data_type.remote_api_order_book_data_source import RemoteAPIOrderBookDataSource
 from hummingbot.market.dolomite.dolomite_api_order_book_data_source import DolomiteAPIOrderBookDataSource
-
-from hummingbot.core.data_type.order_book_message import OrderBookMessageType, DolomiteOrderBookMessage
-from hummingbot.core.data_type.order_book_tracker_entry import DolomiteOrderBookTrackerEntry
+from hummingbot.market.dolomite.dolomite_order_book_tracker_entry import DolomiteOrderBookTrackerEntry
+from hummingbot.core.data_type.order_book_message import OrderBookMessageType
 
 
 class DolomiteOrderBookTracker(OrderBookTracker):
