+++ hummingbot/market/idex/idex_api_order_book_data_source.py
@@ -22,11 +22,9 @@ from hummingbot.core.utils import async_ttl_cache
 from hummingbot.market.idex.idex_active_order_tracker import IDEXActiveOrderTracker
 from hummingbot.market.idex.idex_order_book import IDEXOrderBook
 from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
-from hummingbot.core.data_type.order_book_tracker_entry import (
-    IDEXOrderBookTrackerEntry,
-    OrderBookTrackerEntry
-)
-from hummingbot.core.data_type.order_book_message import IDEXOrderBookMessage
+from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
+from hummingbot.market.idex.idex_order_book_message import IDEXOrderBookMessage
+from hummingbot.market.idex.idex_order_book_tracker_entry import IDEXOrderBookTrackerEntry
 
 IDEX_REST_URL = "https://api.idex.market"
 IDEX_WS_URL = "wss://datastream.idex.market"
