+++ hummingbot/market/idex/idex_order_book_tracker_entry.py
@@ -1,0 +1,23 @@
+from hummingbot.core.data_type.order_book import OrderBook
+from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
+from hummingbot.market.idex.idex_active_order_tracker import IDEXActiveOrderTracker
+
+
+class IDEXOrderBookTrackerEntry(OrderBookTrackerEntry):
+    def __init__(self,
+                 trading_pair: str,
+                 timestamp: float,
+                 order_book: OrderBook,
+                 active_order_tracker: IDEXActiveOrderTracker):
+        self._active_order_tracker = active_order_tracker
+        super(IDEXOrderBookTrackerEntry, self).__init__(trading_pair, timestamp, order_book)
+
+    def __repr__(self) -> str:
+        return (
+            f"IDEXOrderBookTrackerEntry(trading_pair='{self._trading_pair}', timestamp='{self._timestamp}', "
+            f"order_book='{self._order_book}')"
+        )
+
+    @property
+    def active_order_tracker(self) -> IDEXActiveOrderTracker:
+        return self._active_order_tracker
