+++ hummingbot/market/liquid/liquid_order_book_tracker_entry.py
@@ -1,17 +1,18 @@
 #!/usr/bin/env python
 
 from hummingbot.core.data_type.order_book import OrderBook
+from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
 
 
-class OrderBookTrackerEntry:
+class LiquidOrderBookTrackerEntry(OrderBookTrackerEntry):
     def __init__(self, trading_pair: str, timestamp: float, order_book: OrderBook):
         self._trading_pair = trading_pair
         self._timestamp = timestamp
         self._order_book = order_book
 
-    def __repr__(self) -> str:
+    def __repr__(self) -> (str):
         return (
-            f"OrderBookTrackerEntry(trading_pair='{self._trading_pair}', timestamp='{self._timestamp}', "
+            f"LiquidOrderBookTrackerEntry(trading_pair='{self._trading_pair}', timestamp='{self._timestamp}', "
             f"order_book='{self._order_book}')"
         )
