+++ test/integration/test_coinbase_pro_active_order_tracker.py
@@ -13,7 +13,7 @@ from typing import (
 from hummingbot.market.coinbase_pro.coinbase_pro_order_book_tracker import CoinbaseProOrderBookTracker
 from hummingbot.core.data_type.order_book import OrderBook
 from hummingbot.core.data_type.order_book_tracker import OrderBookTrackerDataSourceType
-from hummingbot.core.data_type.order_book_message import CoinbaseProOrderBookMessage
+from hummingbot.market.coinbase_pro.coinbase_pro_order_book_message import CoinbaseProOrderBookMessage
 from hummingbot.core.data_type.order_book_row import OrderBookRow
 
 test_trading_pair = "BTC-USD"
@@ -74,7 +74,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
         open_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(open_message_dict)
         open_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(open_message)
         self.assertEqual(open_ob_row[0], [OrderBookRow(price, open_size, open_sequence)])
-    
+
         # receive change message
         change_size = 50.0
         change_sequence = 2
@@ -90,7 +90,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
             "side": side
         }
         change_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(change_message_dict)
-    
+
         change_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(change_message)
         self.assertEqual(change_ob_row[0], [OrderBookRow(price, change_size, change_sequence)])
 
@@ -110,7 +110,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
             "side": side
         }
         match_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(match_message_dict)
-    
+
         match_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(match_message)
         self.assertEqual(match_ob_row[0], [OrderBookRow(price, change_size - match_size, match_sequence)])
 
@@ -129,7 +129,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
             "remaining_size": "0"
         }
         done_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(done_message_dict)
-    
+
         done_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(done_message)
         self.assertEqual(done_ob_row[0], [OrderBookRow(price, done_size, done_sequence)])
 
@@ -157,7 +157,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
         open_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(open_message_dict)
         open_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(open_message)
         self.assertEqual(open_ob_row[1], [OrderBookRow(price, open_size, open_sequence)])
-        
+
         # receive open sell message to be added to active orders
         order_id_2 = "def"
         side = "sell"
@@ -193,7 +193,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
             "side": side
         }
         change_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(change_message_dict)
-    
+
         change_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(change_message)
         self.assertEqual(change_ob_row[1], [OrderBookRow(price, change_size + open_size_2, change_sequence)])
 
@@ -213,7 +213,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
             "side": side
         }
         match_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(match_message_dict)
-    
+
         match_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(match_message)
         self.assertEqual(match_ob_row[1], [OrderBookRow(price, change_size - match_size + open_size_2, match_sequence)])
 
@@ -232,7 +232,7 @@ class CoinbaseProOrderBookTrackerUnitTest(unittest.TestCase):
             "remaining_size": "0"
         }
         done_message: CoinbaseProOrderBookMessage = test_order_book.diff_message_from_exchange(done_message_dict)
-    
+
         done_ob_row: OrderBookRow = test_active_order_tracker.convert_diff_message_to_order_book_row(done_message)
         self.assertEqual(done_ob_row[1], [OrderBookRow(price, done_size + open_size_2, done_sequence)])
